tools/requirements.txt
python-vlc
colorama